+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/price/dto/PriceCreateForm.java
@@ -1,11 +1,7 @@
 package io.gig.realestate.domain.realestate.price.dto;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author : JAKE
@@ -23,15 +19,15 @@ public class PriceCreateForm {
 
     private double revenueRate;
 
-    private int averageUnitPrice;
+    private double averageUnitPrice;
 
-    private int guaranteePrice;
+    private double guaranteePrice;
 
-    private int rentMonth;
+    private double rentMonth;
 
-    private int management;
+    private double management;
 
-    private int managementExpense;
+    private double managementExpense;
 
     private double totalLndpclArByPyung;
